feat: allow calling callbacks from JS multiple times
## Summary
`CallbackWrapper` inherits after `LongLivedObject`. RN maintains a `shared_ptr` in a global set to it until you call `allowRelease()` on it.
Previously, we always held the wrapper via `weak_ptr` and called `allowRelease()` on it after it was called, freeing the callback.
This PR changes the behaviour to instead store a `shared_ptr` in the returned `std::function` object so that the callback is freed after the `function` object is destroyed instead.

## Changes
- added support for calling callbacks passed from JS to TMs multiple times

## Test Plan
1. Open tester app.
2. Go to TurboModule test suite
3. Verify the "call the passed callback twice" test passes


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Closes #1448

See merge request rnoh/react-native-harmony!1753

Signed-off-by: huangyouhua <1356795413@qq.com>