fix: apply vp scale to transforms correctly
## Summary
Translations are given in different units in RN and ArkUI, so we need to scale them by the vp-to-px factor.
Previously we did that by multiplying certain fields in the transform matrix, which is not correct for compound transformations (and not simple translations).
This PR fixes that.

This was an attempt at fixing the Perspective transform, and while they look better (compare Animated -> Perspective test before and after to Android), something is still off with it.

## Changes
- fixed incorrect transforms in more complex scenarios

## Test Plan
1. Open tester app.
2. Verify no regressions in the Transform test suite.


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Related: #1077

See merge request rnoh/react-native-harmony!1755

Signed-off-by: huangyouhua <1356795413@qq.com>