fix: values for animated props are not set correctly after `display` changes to `none`
## Summary
When setting `display` style prop to `none`, Fabric unmounts the native components in the subtree of the hidden component.

If that component contains views with Animated props, after `display` changes back and Fabric remounts the native components, it will set its properties to values last set from JS, not these set by animated. This for RN <0.73 caused these properties to be set incorrectly.

This PR implements the upstream fix for this issue introduced on iOS in 0.73 by notifying JS of the final animation value in the end callback.

## Changes
- fixed: incorrect values for Animated properties when `display` property of its parent changes to 'none'

## Test Plan
1. Open tester app.
2. Go to Animated Test suite.
3. Verify "maintain translation after toggling display to 'none' and back" test behaves correctly.


## Checklist
- [x] Project compiles and the tester application works.

## Issue Links
Related #1451 -- now behaves the same as on iOS 0.75, but still not fully correctly -- one of the animated views still disappears after navigating back

See merge request rnoh/react-native-harmony!1763

Signed-off-by: huangyouhua <1356795413@qq.com>