fix: getValueAsDouble method return value when m_value is an integer
## Summary
This MR fixes `getValueAsDouble` method in `ValueAnimatedNode` by always casting the value to `double` when `m_value` is a number. Previously, the value was only returned when it was a `double`, however, we should do the same when the number is an integer.

## Changes
- fixed crashes caused by getValueAsDouble method in ValueAnimatedNode

## Test Plan
Tested Animated tests for regressions.

## Checklist
- [x] Project compiles and the tester application works.

See merge request rnoh/react-native-harmony!1799

Signed-off-by: huangyouhua <1356795413@qq.com>