fix: bring back RNOH and custom systrace support
## Why are these changes being made?
After updating RN to version 0.76, the contents of `rn/CMakeLists` have changed significantly, and the configuration responsible for RNOH traces was removed. This MR brings back that configuration and adapts it accordingely.

## Changes
- fixed RNOH Systraces which are broken since the update to RN 0.76 and include libhitrace_ndk.z.so only when necessary

## Test Plan
Run the tester app with profiler and notice there are RNOH specific traces.

## Checklist
- [x] `pnpm verify` passes
- [x] no test fails on `CONCURRENT TESTER` page in tester app
- [x] no test fails on `SEQUENTIAL TESTER` page in tester app

## Issue Links
Closes https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/1497

See merge request rnoh/react-native-harmony!1845

Signed-off-by: huangyouhua <1356795413@qq.com>