chore: add deployment script
## Why are these changes being made?

This MR adds a deployment script for the monorepo. The script publishes react-native-harmony and selected packages, specifically the react-native-core submodule. After publishing, the script creates an MR with a version bump for the react-native repository. If the deployment of react-native-core is selected, another MR is created for the react-native-core package.

## Test Plan
1. run pnpm deploy-packages
2. check if packages are properly deployed

## Checklist
- [x] `pnpm verify` passes
- [x] no test fails on `CONCURRENT TESTER` page in tester app
- [x] no test fails on `SEQUENTIAL TESTER` page in tester app

## Issue Links

See merge request rnoh/react-native-harmony!1852
Closes: https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/1510

Signed-off-by: huangyouhua <1356795413@qq.com>