chore: update MR template
# message auto-generated for no-merge-commit merge:
merge kas/update-mr-template into 0.77.1-rc.1-ohos

chore: update MR template

Created-by: arkadiuszkasprzyk
Commit-by: Arkadiusz
Merged-by: my_performance
Description: ## Why are these changes being made?
Huawei and SWM devs has started recently working on one branch. This MR template aims to unify the workflow for both teams. Also, the current template isn't compatible with the script that generates the changelog.

I changed the header for this section, in order to encourage devs to explain "why" the change is needed, not "what" is being change. I also updated the script that generates changelog, to search for "Changelog" section. I removed the check about performance, since we can't currently realistically verify that for every MR without performance tests or instructions how to do it.

## Changelog


## Test Plan
When creating this MR, I noticed that the current template on the 77 branch wasn't used. I think GitCode uses the template from master branch, and the english template seems to be ignored. On GitLab it's possible to choose one of many templates from the droplist, but it looks it's not possible on the GitCode. I think we will need to manually copy this template on 0.77 branch at least for now.

## Checklist
- [x] `pnpm verify` command has passed successfully.
- [x] No failures occurred in automatic tests on the `CONCURRENT TESTER` page in the tester app.
- [x] No failures occurred in automatic tests on the `SEQUENTIAL TESTER` page in the tester app.
- [x] If the MR includes a breaking change, a review has been requested from a Huawei reviewer.



See merge request: openharmony-sig/ohos_react_native!643