[Fix] image组件 resizeMode 默认值为 strech
Signed-off-by: chenhaoyue <chenhaoyue1@huawei.com>


# message auto-generated for no-merge-commit merge:
merge image-resize-chy into 0.77.1-rc.1-ohos

[Fix] Image组件初始化时设置ResizeMode属性为Stretch

Created-by: qq_41986353
Commit-by: chenhaoyue
Merged-by: ohci1
Description: ## Description
The onPropsChanged function in the Image component compares m_props->resizeMode with props->resizeMode to determine whether to set the ResizeMode property. Since in RN, the resizeMode property of the image component is defaulted to Stretch, and in ArkUI, the resizeMode property of the Image component is defaulted to Cover, when the value set in the component is Stretch, it will not re-render if the default value and the set value are the same. Therefore, the resizeMode property is set to Stretch in the constructor.

## Changes
- fixed: Fix the issue of initializing the default value of the ResizeMode property of the Image component.

## Test Plan
1. run demo
```
import React from 'react';
import { View, Image, StyleSheet } from 'react-native';

const styles = StyleSheet.create({
  container: {
    paddingTop: 50,
  },
  tinyLogo: {
    width: 50,
    height: 50,
    objectFit: 'fill'
  },
  logo: {
    width: 66,
    height: 58,
  },
});

const DisplayAnImage = () => {
  return (
    <View style={styles.container}>
      <Image
        style={styles.tinyLogo}
        source={{
          uri: 'https://reactnative.dev/img/tiny_logo.png',
        }}
      />
    </View>
  );
}

export default DisplayAnImage;
```
2. Expected results
In ArkUIInspector, the objectFit property of the image component is set to ImageFit.Fill.  
![image.png](https://raw.gitcode.com/user-images/assets/4370111/35e3decc-a1e4-400f-b8a1-8b272c98758c/image.png 'image.png')

## Checklist
- [x] Does not involve incompatible changes; if involved, has been reviewed accordingly.
- [ ] Does not impact performance, or performance testing has been conducted without degradation.
- [ ] Complies with the relevant coding standards.
- [ ] Does not involve documentation updates, or the documentation has been updated.
- [x] Meets testability requirements with necessary self-test cases, appropriate logging, or trace information added.
- [x] No illegal file inclusions exist, such as images or code.


See merge request: openharmony-sig/ohos_react_native!858