fix: Specify the image cache directory.
Co-authored-by: 谢代荣<xiedairong@huawei.com>



# message auto-generated for no-merge-commit merge:
merge imageCache into 0.77.1-rc.1-ohos

fix: Specify the image cache directory.

Created-by: xiedrsz
Commit-by: 谢代荣
Merged-by: huangyouhua
Description: ## Why are these changes being made?
If subfolders are added to the cache directory, cleanup failures may occur when the maxSize is reached, as the system might misidentify directories as missing files.
So, it is necessary to specify a cache directory for images in RN and to check whether the current file is a valid file when setting the cache.


## Changelog
- fixed specify the cache directory for RN images in ImageLoaderTurboModule.

## Test Plan
1. Change maxSize to 2 in ImageLoaderTurboModule.
1. Create a folder inside the `rn_image_cache` directory.
1. Call `Image.prefetch(url);` to cache multiple images and check if the program runs correctly.

## Checklist
- [x] `pnpm verify` command has passed successfully.
- [x] No failures occurred in automatic tests on the `CONCURRENT TESTER` page in the tester app.
- [x] No failures occurred in automatic tests on the `SEQUENTIAL TESTER` page in the tester app.
- [x] MR does not include a breaking change or a review has been requested from a Huawei reviewer.


See merge request: openharmony-sig/ohos_react_native!883