fix: DevTool connection is not terminated properly when navigating between RN and native page
Co-authored-by: Wojciech Łoboda<wojciech.loboda@swmansion.com>



# message auto-generated for no-merge-commit merge:
merge close-websocket into 0.77.1-rc.1-ohos

fix: DevTool connection is not terminated properly when navigating between RN and native page

Created-by: wojciech-loboda
Commit-by: Wojciech Łoboda
Merged-by: huangyouhua
Description: <!-- Please remove all comments before submitting to avoid them being included in merge commit message -->
## Why are these changes being made?

Currently, when navigating from a React Native page to a native page and opening DevTools by pressing j in the CLI, DevTools open but are not connected to any target. Instead of showing an unresponsive DevTools window, a warning  "No connected targets" should be displayed.

This issue occurs because the WebSocket connection is not properly terminated when the Inspector is closed. As a result, the connection remains active, and the Inspector target appears to be connectable, even though it shouldn't be.

## Changelog
<!-- List user-visible changes in a concise manner. Use one of the following verbs in past tense:
- added
- fixed
- deprecated
- removed
- optimized
Include [breaking] prefix for breaking changes -->

- fixed WebSocket connection not terminating when Inspector is closed

## Test Plan
- run app that uses native and RN page
- navigate to RN Page
- navigate to native page
- when pressing j in the CLI, warning no connected targets should apear

## Checklist
- [x] `pnpm verify` command has passed successfully.
- [x] No failures occurred in automatic tests on the `CONCURRENT TESTER` page in the tester app.
- [x] No failures occurred in automatic tests on the `SEQUENTIAL TESTER` page in the tester app.
- [x] MR does not include a breaking change or a review has been requested from a Huawei reviewer.


See merge request: openharmony-sig/ohos_react_native!1240