fix: Resolve the incompatibility issue of SurfaceConfig on 77 and 72
Co-authored-by: zheng-jun-feng<949815885@qq.com>



# message auto-generated for no-merge-commit merge:
merge 0.77.1-rc.1-ohos into 0.77.1-rc.1-ohos

fix: Resolve the incompatibility issue of SurfaceConfig on 77 and 72

Created-by: zheng-jun-feng
Commit-by: zheng-jun-feng
Merged-by: huangyouhua
Description: <!-- PLEASE REMOVE COMMENTS BEFORE MERGING TO AVOID INCLUDING THEM IN GIT HISTORY -->
## Description
<!--
Briefly describe the purpose of the MR, for example:
This PR mainly implements the function of remote debugging within the Metro LAN.
-->
Resolve the incompatibility issue of SurfaceConfig on 77 and 72

In our sample project, there is the following usage:
```typescript
RNSurface({
    surfaceConfig: {
        appKey: this.appKey,
    },
   ctx: this.ctx
})
```
At this point, it runs normally on version 72, but an error occurs on version 77:
```typescript
Type '{ appKey: string; }' is not assignable to type 'SurfaceConfig'. Property 'initialProps' is missing in type '{ appKey: string; }' but required in type 'ManagedSurfaceConfig'. <ArkTSCheck>
```
This is because the implementation of `SurfaceConfig` on version 72 is:
```typescript
export interface SurfaceConfig {}

export interface SurfaceConfig2 extends SurfaceConfig {
  appKey: string,
  initialProps: SurfaceProps,
  isSurfaceContentAdaptive?: boolean
}
```
While on version 77, the implementation of `SurfaceConfig` is:
```typescript
export type SurfaceConfig = UnmanagedSurfaceConfig | ManagedSurfaceConfig

export interface ManagedSurfaceConfig {
  appKey: string,
  initialProps: SurfaceProps,
}
```
Therefore, to resolve this incompatibility issue, it is necessary to modify the ManagedSurfaceConfig parameter on 77 to make it optional. 
## Changes
<!--
List main changes in bullet points that can be included in the changelog. Those changes should be written for
application developers and react third-party library authors. Internal changes shouldn't be listed. Please start each bullet point with:
- added
- fixed
- deprecated
- removed

If it is a change to add a new attribute, use [added] prefix, for example:
- [added] FlatList component adds inverted attribute
-->

## Test Plan
<!-- 
How to test this PR. For example:
1. Open the Tester project
2. Execute the npm run start command
...
-->



## Checklist
<!--
Please conduct self inspection according to the following checklist before merging
After completing the self check, fill in "x" in [], For example:
- [x] Does not involve incompatible changes; if involved, has been reviewed accordingly.
-->
- [x] Does not involve incompatible changes; if involved, has been reviewed accordingly.
- [x] Does not impact performance, or performance testing has been conducted without degradation.
- [x] Complies with the relevant coding standards.
- [x] Does not involve documentation updates, or the documentation has been updated.
- [x] Meets testability requirements with necessary self-test cases, appropriate logging, or trace information added.
- [x] No illegal file inclusions exist, such as images or code.


See merge request: openharmony-sig/ohos_react_native!1261