fix: the issue where the Web component fails to initialize when an error occurs during hot reload.
Co-authored-by: 谢代荣<xiedairong@huawei.com>



# message auto-generated for no-merge-commit merge:
merge webview into 0.77.1-rc.1-ohos

fix: the issue where the Web component fails to initialize when an error occurs during hot reload.

Created-by: xiedrsz
Commit-by: 谢代荣
Merged-by: huangyouhua
Description: ## Why are these changes being made?
When using Metro to load the jsBundle, if an error occurs on the JavaScript side, it will trigger the fetchCookies method in httpClient. This method initializes the web engine. If the app has enabled worker threads, the web engine will be initialized in the worker thread, which can cause the initialization of the web engine in the main thread to fail. This PR fixes the issue by initializing the web engine in advance within the httpClient constructor.

## Changelog
- fixed the issue where the Web component fails to initialize when an error occurs during hot reload.

## Test Plan
1. Open the Tester Project.
1. Replace `ContentSlot(this.ctx.getContentForTag(this.tag))` in ContainerView.ets with `Web({ src: 'www.baidu.com', controller: this.controller })`.
1. Check whether the ContainerView component in the CustomNativeComponent test case displays the webpage properly.

## Checklist
- [x] `pnpm verify` command has passed successfully.
- [x] No failures occurred in automatic tests on the `CONCURRENT TESTER` page in the tester app.
- [x] No failures occurred in automatic tests on the `SEQUENTIAL TESTER` page in the tester app.
- [x] MR does not include a breaking change or a review has been requested from a Huawei reviewer.


See merge request: openharmony-sig/ohos_react_native!1502