fix: crash when loading the bundle located under the HSP
Co-authored-by: zheng-jun-feng<949815885@qq.com>



# message auto-generated for no-merge-commit merge:
merge 0.77.1-rc.1-ohos into 0.77.1-rc.1-ohos

fix: crash when loading the bundle located under the HSP

Created-by: zheng-jun-feng
Commit-by: zheng-jun-feng
Merged-by: huangyouhua
Description: <!-- PLEASE REMOVE COMMENTS BEFORE MERGING TO AVOID INCLUDING THEM IN GIT HISTORY -->
## Description
<!--
Briefly describe the purpose of the MR, for example:
This PR mainly implements the function of remote debugging within the Metro LAN.
-->
Determine whether the resourceManager is consistent with RN's resourceManager. If not, use the partner's resourceManager to load the bundle and pass it on to RN.
## Changes
<!--
List main changes in bullet points that can be included in the changelog. Those changes should be written for
application developers and react third-party library authors. Internal changes shouldn't be listed. Please start each bullet point with:
- added
- fixed
- deprecated
- removed

If it is a change to add a new attribute, use [added] prefix, for example:
- [added] FlatList component adds inverted attribute
-->

## Test Plan
<!-- 
How to test this PR. For example:
1. Open the Tester project
2. Execute the npm run start command
...
-->
1. Use to store bundles in HSP
2. Pass in the current hsp module's resourceManager in the ResourceJSBundleProvider
3. Run the application without crashing


## Checklist
<!--
Please conduct self inspection according to the following checklist before merging
After completing the self check, fill in "x" in [], For example:
- [x] Does not involve incompatible changes; if involved, has been reviewed accordingly.
-->
- [ ] Does not involve incompatible changes; if involved, has been reviewed accordingly.
- [ ] Does not impact performance, or performance testing has been conducted without degradation.
- [ ] Complies with the relevant coding standards.
- [ ] Does not involve documentation updates, or the documentation has been updated.
- [ ] Meets testability requirements with necessary self-test cases, appropriate logging, or trace information added.
- [ ] No illegal file inclusions exist, such as images or code.


See merge request: openharmony-sig/ohos_react_native!1630